fixed type error for database

+++ Sketch_32.1_CameraWebServer/hardware_control.h
@@ -82,7 +82,7 @@ void hardware_poll(void);
 
 void hardware_buzz(void);
 
-void send_to_database(bool intruder_status, int face_id, int confidence);
+void send_to_database(bool intruder_status, int face_id, float confidence);
 
 void sendIntruderAlert(void);
